requirements.txt
# Core ML Libraries
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.11.1

# Advanced ML Models
xgboost==1.7.6
lightgbm==4.0.0
optuna==3.3.0

# Deep Learning (optional)
tensorflow==2.13.0
torch==2.0.1

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Dashboard and Web Framework
streamlit==1.25.0
fastapi==0.101.0
uvicorn==0.23.2
pydantic==2.0.3

# Data Processing
openpyxl==3.1.2
xlrd==2.0.1
statsmodels==0.14.0

# API and HTTP
requests==2.31.0
httpx==0.24.1

# Configuration and Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
click==8.1.6

# Logging and Monitoring
loguru==0.7.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Code Quality
black==23.7.0
flake8==6.0.0
isort==5.12.0

# Production and Deployment
gunicorn==21.2.0
docker==6.1.3

# Jupyter and Notebooks
jupyter==1.0.0
ipykernel==6.24.0

# Time Series
prophet==1.1.4

# Feature Engineering
feature-engine==1.6.1

# Model Serialization
joblib==1.3.1
pickle-mixin==1.0.2

# Parallel Processing
dask==2023.7.1

# Database (optional)
sqlalchemy==2.0.19
psycopg2-binary==2.9.7

# Environment Management
python-multipart==0.0.6
jinja2==3.1.2